Remove flycheck-ruby-spring-rubocop-init
To determine whether to use the syntax checker in the current buffer,
we can use :enabled callback instead of
flycheck-ruby-spring-rubocop-init.

And thanks to :next-checker, we don't have to disable ruby-rubocop
explicitly.

When the first cell of :command is non-existent, flycheck disables the
checker, not invoking :enabled callback. So I've replaced "DUMMY" in
:command with "rubocop".